// packages/fx-core/src/component/driver/teamsApp/interfaces/appdefinitions/appDefinition.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT license.

import { AppUser } from "./appUser";
import { AppUserSetting } from "./appUserSetting";
import { AppEnvironment } from "./appEnvironment";
import { AppEnvironmentProperty } from "./appEnvironmentProperty";
import { ConfigurableTab } from "./configurableTab";
import { StaticTab } from "./staticTab";
import { Bot } from "./bot";
import { Connector } from "./connector";
import { SubscriptionOffer } from "./subscriptionOffer";
import { GraphConnector } from "./graphConnector";
import { MessagingExtension } from "./messagingExtension";
import { DefaultGroupCapability } from "./defaultGroupCapability";
import { MeetingExtensionDefinitionItem } from "./meetingExtensionDefinitionItem";
import { AppCategory } from "./appCategory";
import { DisabledScopeItemType } from "./disabledScopeItemType";
import { ActivitiesDefinitionItem } from "./activitiesDefinitionItem";
import { AppItemHostedCapability } from "./appItemHostedCapability";
import { AppItemIndustry } from "./appItemIndustry";
import { AppAuthorizationItem } from "./appAuthorizationItem";
import { LocalizationInfo } from "./localizationInfo";

export interface AppDefinition {
  teamsAppId?: string;
  tenantId?: string;
  ownerAadId?: string;
  userList?: AppUser[];
  userAppSettings?: AppUserSetting[];
  environments?: AppEnvironment[];
  globalEnvironmentProperties?: AppEnvironmentProperty[];
  createdAt?: string;
  updatedAt?: string;
  appId?: string;
  appName?: string;
  appStudioVersion?: string;
  version?: string;
  manifestVersion?: string;
  packageName?: string;
  shortName?: string;
  longName?: string;
  developerName?: string;
  websiteUrl?: string;
  privacyUrl?: string;
  termsOfUseUrl?: string;
  mpnId?: string;
  shortDescription?: string;
  longDescription?: string;
  colorIcon?: string;
  outlineIcon?: string;
  accentColor?: string;
  configurableTabs?: ConfigurableTab[];
  staticTabs?: StaticTab[];
  bots?: Bot[];
  connectors?: Connector[];
  subscriptionOffer?: SubscriptionOffer;
  graphConnector?: GraphConnector;
  messagingExtensions?: MessagingExtension[];
  validDomains?: string[];
  appStudioChecklistChecked?: string[];
  webApplicationInfoId?: string;
  webApplicationInfoResource?: string;
  devicePermissions?: string[];
  applicationPermissions?: string[];
  showLoadingIndicator?: boolean;
  isFullScreen?: boolean;
  isPublic?: boolean;
  defaultInstallScope?: string;
  defaultGroupCapability?: DefaultGroupCapability;
  configurableProperties?: string[];
  meetingExtensionDefinition?: MeetingExtensionDefinitionItem;
  isTeamsOwned?: boolean;
  isBlockable?: boolean;
  isCoreApp?: boolean;
  isPinnable?: boolean;
  isFullTrust?: boolean;
  isTenantConfigurable?: boolean;
  screenshotUrls?: string[];
  keywords?: string[];
  videoUrl?: string;
  categories?: AppCategory[];
  disabledScopes?: DisabledScopeItemType[];
  thirdPartyNoticesUrl?: string;
  isPreinstallable?: boolean;
  activities?: ActivitiesDefinitionItem;
  hostedCapabilities?: AppItemHostedCapability[];
  validTrouterPaths?: string[];
  industries?: AppItemIndustry[];
  authorization?: AppAuthorizationItem;
  defaultBlockUntilAdminAction?: boolean;
  publisherDocsUrl?: string;
  hasPreviewFeature?: boolean;
  localizationInfo?: LocalizationInfo;
  supportsChannelFeatures?: string;
}